Separated age and gender prediction into reusable functions
Modularized the core logic by creating detect_faces() and predict_age_gender() functions.

Cleaned up code for better readability and future reusability (e.g., in Streamlit or GUI apps).

Ensures consistent preprocessing and model inference steps for both webcam and image input modes.
